+++ app/src/main/java/de/matrix44/dedolox/Lfo2Panel.java
@@ -40,19 +40,19 @@ public class Lfo2Panel extends  SynthPanel {
 
     if (event.value1 == MIDIImplementation.CC_LFO2_SPEED) {
       ratePot.blockUpdates(false);
-      ratePot.setValue(event.value2 / 127.0);
+      ratePot.setValue(event.value2 / 127.0f);
       ratePot.blockUpdates(true);
     }
 
     if (event.value1 == MIDIImplementation.CC_LFO2_CUTOFF) {
       filterPot.blockUpdates(false);
-      filterPot.setValue(event.value2 / 127.0);
+      filterPot.setValue(event.value2 / 127.0f);
       filterPot.blockUpdates(true);
     }
 
     if (event.value1 == MIDIImplementation.CC_LFO2_PULSE) {
       pulsePot.blockUpdates(false);
-      pulsePot.setValue(event.value2 / 127.0);
+      pulsePot.setValue(event.value2 / 127.0f);
       pulsePot.blockUpdates(true);
     }
   }
@@ -79,8 +79,8 @@ public class Lfo2Panel extends  SynthPanel {
     ratePot.setControlSteps(25);
     ratePot.setPotListener(new PotView.PotListener() {
       @Override
-      public void onValueChanged(double newVal) {
-        MainAudioThread.getAudioThread().controlChange(0, MIDIImplementation.CC_LFO2_SPEED, (int) (127.0 * newVal));
+      public void onValueChanged(float newVal) {
+        MainAudioThread.getAudioThread().controlChange(0, MIDIImplementation.CC_LFO2_SPEED, (int) (127.0f * newVal));
       }
     });
     addView(ratePot, new FrameLayout.LayoutParams(100, 100));
@@ -88,8 +88,8 @@ public class Lfo2Panel extends  SynthPanel {
     filterPot = new PotView(context);
     filterPot.setPotListener(new PotView.PotListener() {
       @Override
-      public void onValueChanged(double newVal) {
-        MainAudioThread.getAudioThread().controlChange(0, MIDIImplementation.CC_LFO2_CUTOFF, (int) (127.0 * newVal));
+      public void onValueChanged(float newVal) {
+        MainAudioThread.getAudioThread().controlChange(0, MIDIImplementation.CC_LFO2_CUTOFF, (int) (127.0f * newVal));
       }
     });
     addView(filterPot, new FrameLayout.LayoutParams(100, 100));
@@ -97,8 +97,8 @@ public class Lfo2Panel extends  SynthPanel {
     pulsePot = new PotView(context);
     pulsePot.setPotListener(new PotView.PotListener() {
       @Override
-      public void onValueChanged(double newVal) {
-        MainAudioThread.getAudioThread().controlChange(0, MIDIImplementation.CC_LFO2_PULSE, (int) (127.0 * newVal));
+      public void onValueChanged(float newVal) {
+        MainAudioThread.getAudioThread().controlChange(0, MIDIImplementation.CC_LFO2_PULSE, (int) (127.0f * newVal));
       }
     });
     addView(pulsePot, new FrameLayout.LayoutParams(100, 100));
@@ -111,37 +111,37 @@ public class Lfo2Panel extends  SynthPanel {
 
     if (waveSel != null) {
       FrameLayout.LayoutParams params = (FrameLayout.LayoutParams)waveSel.getLayoutParams();
-      params.width      = (int)(0.421875 * width);
-      params.height     = (int)(0.421875 * height);
-      params.leftMargin = (int)(0.060546875 * width);
-      params.topMargin  = (int)(0.069335938 * height);
+      params.width      = (int)(0.421875f * width);
+      params.height     = (int)(0.421875f * height);
+      params.leftMargin = (int)(0.060546875f * width);
+      params.topMargin  = (int)(0.069335938f * height);
       waveSel.setLayoutParams(params);
     }
 
     if (ratePot != null) {
       FrameLayout.LayoutParams params = (FrameLayout.LayoutParams) ratePot.getLayoutParams();
-      params.width      = (int)(0.336914063 * width);
-      params.height     = (int)(0.336914063 * height);
-      params.leftMargin = (int)(0.5546875 * width);
-      params.topMargin  = (int)(0.110351563 * height);
+      params.width      = (int)(0.336914063f * width);
+      params.height     = (int)(0.336914063f * height);
+      params.leftMargin = (int)(0.5546875f * width);
+      params.topMargin  = (int)(0.110351563f * height);
       ratePot.setLayoutParams(params);
     }
 
     if (filterPot != null) {
       FrameLayout.LayoutParams params = (FrameLayout.LayoutParams) filterPot.getLayoutParams();
-      params.width      = (int)(0.336914063 * width);
-      params.height     = (int)(0.336914063 * height);
-      params.leftMargin = (int)(0.5546875 * width);
-      params.topMargin  = (int)(0.5546875 * height);
+      params.width      = (int)(0.336914063f * width);
+      params.height     = (int)(0.336914063f * height);
+      params.leftMargin = (int)(0.5546875f * width);
+      params.topMargin  = (int)(0.5546875f * height);
       filterPot.setLayoutParams(params);
     }
 
     if (pulsePot != null) {
       FrameLayout.LayoutParams params = (FrameLayout.LayoutParams) pulsePot.getLayoutParams();
-      params.width      = (int)(0.336914063 * width);
-      params.height     = (int)(0.336914063 * height);
-      params.leftMargin = (int)(0.109375 * width);
-      params.topMargin  = (int)(0.5546875 * height);
+      params.width      = (int)(0.336914063f * width);
+      params.height     = (int)(0.336914063f * height);
+      params.leftMargin = (int)(0.109375f * width);
+      params.topMargin  = (int)(0.5546875f * height);
       pulsePot.setLayoutParams(params);
     }
